model_serving/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML Frameworks
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
onnx==1.15.0
onnxruntime-gpu==1.16.3

# NVIDIA Libraries
nvidia-tensorrt==8.6.1
nvidia-cuda-runtime-cu12==12.1.0
nvidia-cudnn-cu12==8.9.7

# vLLM for LLM Serving
vllm==0.2.7
ray[default]==2.9.1

# Distributed Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.22.0
opentelemetry-exporter-otlp==1.22.0

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
py-cpuinfo==9.0.0
psutil==5.9.7
pynvml==11.5.0

# Data Processing
numpy==1.26.3
pillow==10.2.0
opencv-python-headless==4.9.0.80

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Logging
structlog==24.1.0
python-json-logger==2.0.7
colorlog==6.8.0

# Database (for model metadata)
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Caching
redis==5.0.1
hiredis==2.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
locust==2.20.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.24.0

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0
typer==0.9.0
